+++ src/main/java/Pix/Event.java
@@ -1,9 +1,11 @@
+package Pix;
+
 public class Event extends Task {
     private String eventDate;
     /**
-     * Constructor for the Deadline task.
-     * @param name Name of the Task.
-     * @param eventDate Date of the Event.
+     * Constructor for the Pix.Deadline task.
+     * @param name Name of the Pix.Task.
+     * @param eventDate Date of the Pix.Event.
      */
     public Event(String name, String eventDate) {
         super(name);
