+++ src/main/java/Pix/ToDo.java
@@ -1,7 +1,9 @@
+package Pix;
+
 public class ToDo extends Task {
     /**
-     * Constructor for the ToDo task.
-     * @param name Name of the Task.
+     * Constructor for the Pix.ToDo task.
+     * @param name Name of the Pix.Task.
      */
     public ToDo(String name) {
         super(name);
