+++ src/main/java/Pix/exception/PixInvalidTaskException.java
@@ -1,0 +1,9 @@
+package Pix.exception;
+
+public class PixInvalidTaskException extends PixException{
+
+    @Override
+    public String getMessage() {
+        return "(ಠ ∩ಠ) The format to enter the specified task is wrong";
+    }
+}
