+++ src/main/java/Pix/exception/PixUnknownCommandException.java
@@ -1,0 +1,12 @@
+package Pix.exception;
+
+/**
+ * Triggers when the user inputs an unknown command.
+ */
+public class PixUnknownCommandException extends PixException {
+
+    @Override
+    public String getMessage() {
+        return "(ಠ ∩ಠ) Unknown Command: Do you even know what you're typing..?";
+    }
+}
